Finish moving readline stuff to freeciv::server
History is now working correctly.
